feat: добавил Interceptor
AuthenticationInterceptor для проверки куки и возврата сессии, чтобы сделать это один раз и сохранить юзера, а не каждый раз искать его по новой

+++ src/main/java/org/pet/project/controller/WeatherController.java
@@ -8,7 +8,6 @@ import org.pet.project.model.dto.api.UserWeatherCardDto;
 import org.pet.project.model.dto.api.entity.Coordinates;
 import org.pet.project.model.entity.User;
 import org.pet.project.service.LocationService;
-import org.pet.project.service.UserSessionCheckService;
 import org.pet.project.service.WeatherApiService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -25,25 +24,25 @@ import java.util.*;
 @RequiredArgsConstructor
 public class WeatherController {
 
-    private final UserSessionCheckService sessionCheckService;
     private final WeatherApiService weatherApiService;
     private final LocationService locationService;
 
     @GetMapping("/")
     public String getWeather(HttpServletRequest req, Model model) {
 
-        Optional<User> optionalUser = sessionCheckService.getAuthenticatedUser(req, model);
+        User user = (User) req.getAttribute("user");
 
-        if (optionalUser.isEmpty()) {
+        if (user == null) {
             return "main-page";
         }
 
-        User user = optionalUser.get();
 
-        List<UserWeatherCardDto> userLocationCards = user.getLocations().stream()
-                .map(weatherApiService::findWeatherByLocation).toList().reversed();
+        List<UserWeatherCardDto> userLocationCards = user.getLocations()
+                .stream()
+                .map(weatherApiService::findWeatherByLocation)
+                .toList()
+                .reversed();
 
-        model.addAttribute("isAuthenticated", true);
         model.addAttribute("login", user.getLogin());
         model.addAttribute("locations", userLocationCards);
         return "main-page";
@@ -52,8 +51,10 @@ public class WeatherController {
     @PostMapping("/delete")
     public String deleteWeatherCard(@ModelAttribute Coordinates coordinates, HttpServletRequest req, Model model) {
 
-        User user = sessionCheckService.getAuthenticatedUser(req, model)
-                .orElseThrow(() -> new CookieNotFoundException("User, which deletes location, is not found"));
+        User user = (User) req.getAttribute("user");
+        if (user == null) {
+            throw new CookieNotFoundException("User, which deletes location, is not found");
+        }
 
         locationService.deleteLocationByUser(user, coordinates.getLat(), coordinates.getLon());
